requirements.txt
fastapi
mangum
python-multipart
python-dotenv
cryptography
python-jose
bcrypt